Split nsys in multiple files into subpackage
