+++ reprospect/tools/nsys/report.py
@@ -1,24 +1,16 @@
 import dataclasses
 import functools
-import json
 import logging
-import os
 import pathlib
 import re
-import shlex
-import shutil
 import sqlite3
-import subprocess
 import sys
 import typing
 
-import attrs
-import blake3
 import pandas
 import rich.tree
 
-from reprospect.tools import cacher
-from reprospect.utils import ldd, rich_helpers
+from reprospect.utils import rich_helpers
 
 if sys.version_info >= (3, 12):
     from typing import override
@@ -29,163 +21,6 @@ if sys.version_info >= (3, 11):
     from typing import Self
 else:
     from typing_extensions import Self
-
-@attrs.define(frozen=True, slots=True, kw_only=True)
-class Command: # pylint: disable=too-many-instance-attributes, duplicate-code
-    """
-    Run a ``nsys`` command line.
-    """
-    executable: str | pathlib.Path
-    """Executable to run."""
-    output: pathlib.Path
-    """Report file."""
-    opts: tuple[str, ...] = ()
-    """Options that do not involve paths."""
-    nvtx_capture: str | None = None
-    """NVTX capture string."""
-    capture_range_end: str = 'stop'
-    """NVTX capture range end."""
-    args: tuple[str | pathlib.Path, ...] | None = None
-    """Arguments to pass to the executable."""
-    env: typing.Mapping[str, str] | None = None
-    """Mapping used to update the environment before running, see :py:meth:`run`."""
-
-    cmd: tuple[str | pathlib.Path, ...] = attrs.field(init=False)
-
-    def __attrs_post_init__(self) -> None:
-        """
-        Enrich :py:attr:`opts` and build :py:attr:`cmd`.
-        """
-        if self.output.suffix != '.nsys-rep':
-            object.__setattr__(self, 'output', self.output.parent / f'{self.output.name}.nsys-rep')
-
-        # Disable collecting CPU samples.
-        opts: tuple[str, ...] = self.opts + (
-            '--sample=none',
-            '--backtrace=none',
-            '--cpuctxsw=none',
-        )
-
-        # We want to start data collection when the first NVTX range is met.
-        # This reduces the amount of data collected (and makes things faster).
-        if self.nvtx_capture is not None:
-            match self.nvtx_capture:
-                case '*':
-                    pass
-                case _:
-                    opts += (
-                        '--capture-range=nvtx',
-                        f'--capture-range-end={self.capture_range_end}',
-                        f'--nvtx-capture={self.nvtx_capture}',
-                    )
-            opts += ('--trace=nvtx,cuda',)
-        else:
-            opts += ('--trace=cuda',)
-
-        object.__setattr__(self, 'opts', opts)
-
-        # Build the final full command.
-        object.__setattr__(self, 'cmd', (
-            'nsys', 'profile',
-            *self.opts,
-            '--force-overwrite=true', '-o', self.output,
-            self.executable,
-            *(self.args or ()),
-        ))
-
-    def run(self, *,
-        cwd: pathlib.Path | None = None,
-        env: typing.MutableMapping[str, str] | None = None,
-    ) -> int:
-        if (self.nvtx_capture is not None or self.env is not None) and env is None:
-            env = os.environ.copy()
-
-        # For '--capture-range=nvtx' to accept our custom strings, we need to allow unregistered
-        # strings to be considered.
-        # See https://docs.nvidia.com/nsight-systems/UserGuide/index.html#example-interactive-cli-command-sequences.
-        if self.nvtx_capture:
-            assert env is not None
-            env['NSYS_NVTX_PROFILER_REGISTER_ONLY'] = '0'
-
-        if self.env:
-            assert env is not None
-            env.update(self.env)
-
-        self.output.unlink(missing_ok=True)
-        return subprocess.check_call(args=self.cmd, env=env, cwd=cwd)
-
-@dataclasses.dataclass(frozen=True, slots=True)
-class Session:
-    """
-    `Nsight Systems` session interface.
-    """
-    command: Command
-
-    def run(
-        self,
-        cwd: pathlib.Path | None = None,
-        env: typing.MutableMapping[str, str] | None = None,
-    ) -> None:
-        """
-        Run ``nsys`` using :py:attr:`command`.
-        """
-        logging.info(f"Launching 'nsys' with {self.command.cmd}.")
-        self.command.run(cwd=cwd, env=env)
-
-    def export_to_sqlite(
-        self,
-        cwd: pathlib.Path | None = None,
-    ) -> pathlib.Path:
-        """
-        Export report to ``.sqlite``.
-        """
-        output_file_sqlite = self.command.output.with_suffix('.sqlite')
-
-        cmd: tuple[str | pathlib.Path, ...] = (
-            'nsys', 'export',
-            '--type', 'sqlite',
-            f'--output={output_file_sqlite}',
-            self.command.output,
-        )
-
-        logging.info(f"Exporting to 'sqlite' with {cmd}.")
-        output_file_sqlite.unlink(missing_ok=True)
-        subprocess.check_call(cmd, cwd=cwd)
-
-        return output_file_sqlite
-
-    def extract_statistical_report(
-        self,
-        report: str = 'cuda_api_sum',
-        filter_nvtx: str | None = None,
-        cwd: pathlib.Path | None = None,
-    ) -> pandas.DataFrame:
-        """
-        Extract `report`, filtering the database with `filter_nvtx`.
-        """
-        output_file_sqlite = self.command.output.with_suffix('.sqlite')
-
-        cmd: tuple[str | pathlib.Path, ...] = (
-            'nsys', 'stats',
-            f'--output={self.command.output.parent / self.command.output.stem}',
-            f'--report={report}',
-            '--format=csv',
-            '--timeunit=usec',
-            *(('--filter-nvtx=' + filter_nvtx,) if filter_nvtx else ()),
-            output_file_sqlite,
-        )
-
-        # 'nsys stats' will output to a file whose name follows the convention
-        #    <basename>_<analysis&args>.<output_format>
-        suffix = '_nvtx=' + filter_nvtx.replace('/', '-') if filter_nvtx else ''
-        output_file_csv = self.command.output.parent / f'{self.command.output.stem}_{report}{suffix}.csv'
-
-        logging.info(f'Extracting statistical report \'{report}\' from {output_file_sqlite} with {cmd}.')
-        output_file_csv.unlink(missing_ok=True)
-        subprocess.check_call(cmd, cwd=cwd)
-        assert output_file_csv.is_file()
-
-        return pandas.read_csv(output_file_csv)
 
 @dataclasses.dataclass(frozen=True, slots=True)
 class ReportPatternSelector:
@@ -429,111 +264,3 @@ def strip_cuda_api_suffix(call: str) -> str:
     Strip suffix like `_v10000` or `_ptsz` from a CUDA API `call`.
     """
     return call.split('_', maxsplit=1)[0]
-
-class Cacher(cacher.Cacher):
-    """
-    Cacher tailored to ``nsys`` results.
-
-    ``nsys`` runs require quite some time to acquire results.
-
-    On a cache hit, the cacher will serve:
-
-    - ``.nsys-rep`` file
-    - ``.sqlite`` file
-
-    On a cache miss, ``nsys`` is launched and the cache entry populated accordingly.
-
-    .. note::
-
-        It is assumed that hashing is faster than running ``nsys`` itself.
-
-    .. warning::
-
-        The cache should not be shared between machines, since there may be differences between machines
-        that influence the results but are not included in the hashing.
-    """
-    TABLE: typing.ClassVar[str] = 'nsys'
-
-    def __init__(self, *, directory: str | pathlib.Path | None = None):
-        super().__init__(directory=directory or (pathlib.Path(os.environ['HOME']) / '.nsys-cache'))
-
-    def hash_impl(self, *, command: Command) -> blake3.blake3:
-        """
-        Hash based on:
-
-        * ``nsys`` version
-        * ``nsys`` options (but not the output files)
-        * executable content
-        * executable arguments
-        * linked libraries
-        * environment
-        """
-        hasher = blake3.blake3() # pylint: disable=not-callable
-
-        hasher.update(subprocess.check_output(('nsys', '--version')))
-
-        if command.opts:
-            hasher.update(shlex.join(command.opts).encode())
-
-        hasher.update_mmap(command.executable)
-
-        if command.args:
-            hasher.update(shlex.join(map(str, command.args)).encode())
-
-        if command.env:
-            hasher.update(json.dumps(command.env).encode())
-
-        for lib in sorted(ldd.get_shared_dependencies(file=command.executable)):
-            hasher.update_mmap(lib)
-
-        return hasher
-
-    @override
-    def hash(self, **kwargs) -> blake3.blake3:
-        return self.hash_impl(command=kwargs['command'])
-
-    @override
-    def populate(self, directory: pathlib.Path, **kwargs) -> None:
-        """
-        When there is a cache miss, call :py:meth:`reprospect.tools.nsys.Session.run`.
-        Fill the `directory` with the artifacts.
-        """
-        command = kwargs.pop('command')
-
-        Session(command=command).run(**kwargs)
-
-        shutil.copy(dst=directory, src=command.output)
-
-    def run(self, command: Command, **kwargs) -> cacher.Cacher.Entry:
-        """
-        On a cache hit, copy files from the cache entry.
-        """
-        entry = self.get(command=command, **kwargs)
-
-        if entry.cached:
-            shutil.copytree(entry.directory, command.output.parent, dirs_exist_ok=True)
-
-        return entry
-
-    @staticmethod
-    def export_to_sqlite(
-        command: Command,
-        entry: cacher.Cacher.Entry,
-        **kwargs,
-    ) -> pathlib.Path:
-        """
-        Export report to ``.sqlite``.
-        """
-        output_file_sqlite = command.output.with_suffix('.sqlite')
-
-        cached = entry.directory / output_file_sqlite.name
-
-        if cached.is_file():
-            logging.info(f'Serving {output_file_sqlite} from the cache entry {entry}.')
-            shutil.copyfile(src=cached, dst=output_file_sqlite)
-        else:
-            logging.info(f'Populating the cache entry {entry} with {output_file_sqlite} from the cache entry {entry}.')
-            Session(command=command).export_to_sqlite(**kwargs)
-            shutil.copyfile(src=output_file_sqlite, dst=cached)
-
-        return output_file_sqlite
