+++ setup.py
@@ -54,7 +54,9 @@ def enable_mypyc(dist: Distribution) -> None:
             'reprospect/tools/ncu/metrics.py',
             'reprospect/tools/ncu/report.py',
             'reprospect/tools/ncu/session.py',
-            'reprospect/tools/nsys.py',
+            'reprospect/tools/nsys/cacher.py',
+            'reprospect/tools/nsys/report.py',
+            'reprospect/tools/nsys/session.py',
             'reprospect/tools/sass/controlflow.py',
             'reprospect/tools/sass/decode.py',
             'reprospect/utils/cmake.py',
